src/main/java/com/axreng/backend/application/idgenerator/RandomIdString.java
package com.axreng.backend.application.idgenerator;

import java.security.SecureRandom;
import java.util.Locale;
import java.util.Objects;
import java.util.Random;

public class RandomIdString {

    public static final String UPPER = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";

    public static final String LOWER = UPPER.toLowerCase(Locale.ROOT);

    public static final String DIGITS = "0123456789";

    public static final String ALPHANUMERIC = UPPER + LOWER + DIGITS;

    private final Random random;

    private final char[] symbols;

    private final char[] buf;

    public RandomIdString(int length, Random random, String symbols) {
        if (length < 1 || symbols.length() < 2) {
            throw new IllegalArgumentException();
        }
        this.random = Objects.requireNonNull(random);
        this.symbols = symbols.toCharArray();
        this.buf = new char[length];
    }

    public String nextString() {
        for (int idx = 0; idx < buf.length; ++idx)
            buf[idx] = symbols[random.nextInt(symbols.length)];
        return new String(buf);
    }

    public RandomIdString(int length, Random random) {
        this(length, random, ALPHANUMERIC);
    }

    public RandomIdString(int length) {
        this(length, new SecureRandom());
    }

    public RandomIdString() {
        this(21);
    }

}